Rename the sentinel package to sentinel2.
Stactools will have packages for other sentinel datasets like Sentinel
2 SAR, so we should differentiate between the programs.
