+++ README.md
@@ -77,6 +77,7 @@ See the [documentation page](https://stactools.readthedocs.io/en/latest/) for th
 | `stactools_corine`            | Methods and commands for working with CORINE Land Cover data                 |
 | `stactools_planet`            | Methods and commands for working with planet data                |
 | `stactools_landsat`           | Methods and commands for working with landsat data        |
+| `stactools_sentinel2`           | Methods and commands for working with Sentinel 2 data        |
 | `stactools_browse`            | Contains a command for launching stac-browser against a local STAC |
 
 Subpackages are symlinked to the `stactools` directory in this repo to allow them to be importable for python running at the top level directory of the repository clone.
