+++ docker/Dockerfile
@@ -74,9 +74,9 @@ RUN pip install -r /tmp/naip/requirements.txt
 COPY stactools_planet/requirements.txt /tmp/planet/requirements.txt
 RUN pip install -r /tmp/planet/requirements.txt
 
-# Sentinel
-COPY stactools_sentinel/requirements.txt /tmp/sentinel/requirements.txt
-RUN pip install -r /tmp/sentinel/requirements.txt
+# Sentinel 2
+COPY stactools_sentinel2/requirements.txt /tmp/sentinel2/requirements.txt
+RUN pip install -r /tmp/sentinel2/requirements.txt
 
 # Jupyter
 RUN pip install jupyter==1.0.0
