+++ scripts/env
@@ -9,7 +9,7 @@ STACTOOLS_SUBPACKAGE_DIRS=(
     "stactools_planet"
     "stactools_naip"
     "stactools_browse"
-    "stactools_sentinel"
+    "stactools_sentinel2"
 );
 
 STACTOOLS_COVERAGE_DIRS=(
@@ -21,5 +21,5 @@ STACTOOLS_COVERAGE_DIRS=(
     "stactools_planet/stactools/planet"
     "stactools_naip/stactools/naip"
     "stactools_browse/stactools/browse"
-    "stactools_sentinel/stactools/sentinel"
+    "stactools_sentinel2/stactools/sentinel2"
 );
