+++ setup.py
@@ -34,7 +34,7 @@ subpackages = [
     Subpackage('landsat', is_extra=True),
     Subpackage('planet', is_extra=True),
     Subpackage('browse', is_extra=True),
-    Subpackage('sentinel', is_extra=True)
+    Subpackage('sentinel2', is_extra=True)
 ]
 
 here = os.path.abspath(os.path.dirname(__file__))
