--- stactools/sentinel
@@ -1,1 +1,0 @@
-../stactools_sentinel/stactools/sentinel
