+++ stactools/sentinel2
@@ -1,0 +1,1 @@
+../stactools_sentinel2/stactools/sentinel2
