--- stactools_sentinel/README.md
@@ -1,5 +1,0 @@
-# stactools sentinel
-
-A subpackage of stactools for working with Sentinel data.
-
-__WORK IN PROGRESS__. There is no functionality in this libary yet.
