--- stactools_sentinel/stactools/sentinel/__init__.py
@@ -1,6 +1,0 @@
-def register_plugin(registry):
-    # Register subcommands
-
-    from stactools.sentinel import commands
-
-    registry.register_subcommand(commands.create_sentinel_command)
