+++ stactools_sentinel2/README.md
@@ -1,0 +1,3 @@
+# stactools sentinel2
+
+A subpackage of stactools for working with Sentinel data.
