+++ stactools_sentinel2/stactools/sentinel2/__init__.py
@@ -1,0 +1,6 @@
+def register_plugin(registry):
+    # Register subcommands
+
+    from stactools.sentinel2 import commands
+
+    registry.register_subcommand(commands.create_sentinel2_command)
