+++ stactools_sentinel2/stactools/sentinel2/constants.py
